functions/legislative-analysis-func/requirements.txt
functions-framework==3.*
# Let pip find the correct, compatible versions for Google libraries
google-cloud-aiplatform
google-cloud-storage
google-cloud-firestore
google-auth